+++ sdk/metric/meter.go
@@ -27,8 +27,9 @@ var ErrInstrumentName = errors.New("invalid instrument name")
 type meter struct {
 	embedded.Meter
 
-	scope instrumentation.Scope
-	pipes pipelines
+	scope        instrumentation.Scope
+	pipes        pipelines
+	allowDupKeys bool
 
 	int64Insts             *cacheWithErr[instID, *int64Inst]
 	float64Insts           *cacheWithErr[instID, *float64Inst]
@@ -39,7 +40,7 @@ type meter struct {
 	float64Resolver resolver[float64]
 }
 
-func newMeter(s instrumentation.Scope, p pipelines) *meter {
+func newMeter(s instrumentation.Scope, p pipelines, allowDupKeys bool) *meter {
 	// viewCache ensures instrument conflicts, including number conflicts, this
 	// meter is asked to create are logged to the user.
 	var viewCache cache[string, instID]
@@ -58,6 +59,7 @@ func newMeter(s instrumentation.Scope, p pipelines) *meter {
 		float64ObservableInsts: &float64ObservableInsts,
 		int64Resolver:          newResolver[int64](p, &viewCache),
 		float64Resolver:        newResolver[float64](p, &viewCache),
+		allowDupKeys:           allowDupKeys,
 	}
 }
 
@@ -163,7 +165,7 @@ func (m *meter) int64ObservableInstrument(
 			// is not part of the pipeline.
 			insert.pipeline.addInt64Measure(inst.observableID, in)
 			for _, cback := range callbacks {
-				inst := int64Observer{measures: in}
+				inst := int64Observer{measures: in, allowDupKeys: m.allowDupKeys}
 				fn := cback
 				insert.addCallback(func(ctx context.Context) error { return fn(ctx, inst) })
 			}
@@ -346,7 +348,7 @@ func (m *meter) float64ObservableInstrument(
 			// is not part of the pipeline.
 			insert.pipeline.addFloat64Measure(inst.observableID, in)
 			for _, cback := range callbacks {
-				inst := float64Observer{measures: in}
+				inst := float64Observer{measures: in, allowDupKeys: m.allowDupKeys}
 				fn := cback
 				insert.addCallback(func(ctx context.Context) error { return fn(ctx, inst) })
 			}
@@ -523,7 +525,7 @@ func (m *meter) RegisterCallback(f metric.Callback, insts ...metric.Observable) 
 
 	unregs := make([]func(), len(m.pipes))
 	for ix, pipe := range m.pipes {
-		reg := newObserver(pipe)
+		reg := newObserver(pipe, m.allowDupKeys)
 		for _, inst := range validInstruments {
 			switch o := inst.(type) {
 			case int64Observable:
@@ -544,16 +546,18 @@ func (m *meter) RegisterCallback(f metric.Callback, insts ...metric.Observable) 
 type observer struct {
 	embedded.Observer
 
-	pipe    *pipeline
-	float64 map[observableID[float64]]struct{}
-	int64   map[observableID[int64]]struct{}
+	pipe         *pipeline
+	float64      map[observableID[float64]]struct{}
+	int64        map[observableID[int64]]struct{}
+	allowDupKeys bool
 }
 
-func newObserver(p *pipeline) observer {
+func newObserver(p *pipeline, allowDupKeys bool) observer {
 	return observer{
-		pipe:    p,
-		float64: make(map[observableID[float64]]struct{}),
-		int64:   make(map[observableID[int64]]struct{}),
+		pipe:         p,
+		float64:      make(map[observableID[float64]]struct{}),
+		int64:        make(map[observableID[int64]]struct{}),
+		allowDupKeys: allowDupKeys,
 	}
 }
 
@@ -594,7 +598,7 @@ func (r observer) ObserveFloat64(o metric.Float64Observable, v float64, opts ...
 	}
 	c := metric.NewObserveConfig(opts)
 	rawKVs := extractRawKVs(opts)
-	set := resolveAttributes(c.Attributes(), rawKVs)
+	set := resolveAttributes(c.Attributes(), rawKVs, r.allowDupKeys)
 	// Access to r.pipe.float64Measure is already guarded by a lock in pipeline.produce.
 	// TODO (#5946): Refactor pipeline and observable measures.
 	measures := r.pipe.float64Measures[oImpl.observableID]
@@ -627,7 +631,7 @@ func (r observer) ObserveInt64(o metric.Int64Observable, v int64, opts ...metric
 	}
 	c := metric.NewObserveConfig(opts)
 	rawKVs := extractRawKVs(opts)
-	set := resolveAttributes(c.Attributes(), rawKVs)
+	set := resolveAttributes(c.Attributes(), rawKVs, r.allowDupKeys)
 	// Access to r.pipe.int64Measures is already guarded b a lock in pipeline.produce.
 	// TODO (#5946): Refactor pipeline and observable measures.
 	measures := r.pipe.int64Measures[oImpl.observableID]
@@ -695,7 +699,7 @@ func (p int64InstProvider) lookup(
 		Kind:        kind,
 	}, func() (*int64Inst, error) {
 		aggs, err := p.aggs(kind, name, desc, u, allowedKeys)
-		return &int64Inst{measures: aggs}, err
+		return &int64Inst{measures: aggs, allowDupKeys: p.allowDupKeys}, err
 	})
 }
 
@@ -712,7 +716,7 @@ func (p int64InstProvider) lookupHistogram(
 		Kind:        InstrumentKindHistogram,
 	}, func() (*int64Inst, error) {
 		aggs, err := p.histogramAggs(name, cfg, allowedKeys)
-		return &int64Inst{measures: aggs}, err
+		return &int64Inst{measures: aggs, allowDupKeys: p.allowDupKeys}, err
 	})
 }
 
@@ -769,7 +773,7 @@ func (p float64InstProvider) lookup(
 		Kind:        kind,
 	}, func() (*float64Inst, error) {
 		aggs, err := p.aggs(kind, name, desc, u, allowedKeys)
-		return &float64Inst{measures: aggs}, err
+		return &float64Inst{measures: aggs, allowDupKeys: p.allowDupKeys}, err
 	})
 }
 
@@ -786,30 +790,32 @@ func (p float64InstProvider) lookupHistogram(
 		Kind:        InstrumentKindHistogram,
 	}, func() (*float64Inst, error) {
 		aggs, err := p.histogramAggs(name, cfg, allowedKeys)
-		return &float64Inst{measures: aggs}, err
+		return &float64Inst{measures: aggs, allowDupKeys: p.allowDupKeys}, err
 	})
 }
 
 type int64Observer struct {
 	embedded.Int64Observer
 	measures[int64]
+	allowDupKeys bool
 }
 
 func (o int64Observer) Observe(val int64, opts ...metric.ObserveOption) {
 	c := metric.NewObserveConfig(opts)
 	rawKVs := extractRawKVs(opts)
-	o.observe(val, resolveAttributes(c.Attributes(), rawKVs))
+	o.observe(val, resolveAttributes(c.Attributes(), rawKVs, o.allowDupKeys))
 }
 
 type float64Observer struct {
 	embedded.Float64Observer
 	measures[float64]
+	allowDupKeys bool
 }
 
 func (o float64Observer) Observe(val float64, opts ...metric.ObserveOption) {
 	c := metric.NewObserveConfig(opts)
 	rawKVs := extractRawKVs(opts)
-	o.observe(val, resolveAttributes(c.Attributes(), rawKVs))
+	o.observe(val, resolveAttributes(c.Attributes(), rawKVs, o.allowDupKeys))
 }
 
 func defaultAttributes[T any](opts []T) []attribute.Key {